moves files - App to components - fetch and registerService to libraries - updates index.js with correct paths
