+++ client/src/libraries/registerServiceWorker.js
